% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_types.R
\name{Data.frame}
\alias{Data.frame}
\title{Dataframe type}
\usage{
Data.frame(
  columns = list(),
  default_type = typed::Any(),
  select = "any_of",
  only_cols = FALSE,
  empty_ok = TRUE,
  null_ok = FALSE,
  anyNA = FALSE,
  ...
)
}
\arguments{
\item{columns}{List where each element is a column where it must have a declared type, ex list(a = typed::Double())}

\item{default_type}{On elements not declared on types, this will be the default one}

\item{select}{if "all_of" all the elements declared on types must exists}

\item{only_cols}{If must exists all and only the columns declared on types}

\item{empty_ok}{If is acceptable to be empty}

\item{null_ok}{Parsed to typed::Data.frame}

\item{anyNA}{Parsed to typed::Data.frame}

\item{...}{Parsed to assertion_factory}
}
\value{
Assertion for Dataframe
}
\description{
typed Dataframe with custom types
}
