% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geotypes-package.R
\docType{package}
\name{geotypes-package}
\alias{geotypes}
\alias{geotypes-package}
\title{geotypes: Basic static types for spatial}
\description{
Basic static types for spatial.
}
\author{
\strong{Maintainer}: Felipe Matas \email{felipematas@yahoo.com}

}
\keyword{internal}
