% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_sf}
\alias{sf_sf}
\title{SF: SF Object}
\usage{
sf_sf(df_opts = list(), active_column = NULL, active_types = NULL, ...)
}
\arguments{
\item{df_opts}{SF object is a dataframe too, this is a list where they will be parsed to the Data.frame type}

\item{active_column}{Name of the active column}

\item{active_types}{Possible types for the active column}

\item{...}{Parsed to assertion_factory}
}
\value{
Assertion for sf
}
\description{
typed for SF object from SF
}
