% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_sfc}
\alias{sf_sfc}
\title{SF: Geometry Column Type}
\usage{
sf_sfc(types = NULL, only_valid = NULL, ...)
}
\arguments{
\item{types}{Vector with possible acceptable geometry types}

\item{only_valid}{TRUE to accept sfc only with valid geometries}

\item{...}{Parsed to assertion_factory}
}
\value{
Assertion for sfg
}
\description{
typed for geometry column types from SF
}
