% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_sfg}
\alias{sf_sfg}
\title{SF: Geometry Type}
\usage{
sf_sfg(types = NULL, ...)
}
\arguments{
\item{types}{Vector with possible acceptable geometry types}

\item{...}{Parsed to assertion_factory}
}
\value{
Assertion for sfg
}
\description{
typed for geometry types from SF
}
