% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfnetworks.R
\name{sfnetworks_sfnetwork}
\alias{sfnetworks_sfnetwork}
\title{SFnetwork type from sfnetworks}
\usage{
sfnetworks_sfnetwork(null_ok = FALSE, ...)
}
\arguments{
\item{null_ok}{If NULL is acceptable}

\item{...}{Parsed to assertion_factory}
}
\value{
Assertion for sfnetwork
}
\description{
typed for sfnetwork object from sfnetworks
}
