% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_types.R
\name{Data.frame}
\alias{Data.frame}
\title{Dataframe type}
\usage{
Data.frame(
  columns = list(),
  default_type = typed::Any(),
  select = "any_of",
  empty_ok = TRUE,
  null_ok = FALSE,
  anyNA = FALSE,
  ...
)
}
\arguments{
\item{columns}{List where each element is a column where
it must have a declared type, ex list(a = typed::Double())}

\item{default_type}{On elements not declared
on types, this will be the default one}

\item{select}{Set the relations of the declared columns and the existing ones
"any_of": Can exists any columns on the dataframe
"only_from": Can only exists declared columns
(don't need to exists all of them), empty df is valid
"all_of": Must exists at least all the declared columns
"exclusively": Must exists and only the declared columns}

\item{empty_ok}{If is acceptable to be empty}

\item{null_ok}{Parsed to typed::Data.frame}

\item{anyNA}{Parsed to typed::Data.frame}

\item{...}{Parsed to assertion_factory}
}
\value{
Assertion for Dataframe
}
\description{
typed Dataframe with custom types
}
