% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.R
\name{sf_sf}
\alias{sf_sf}
\title{SF: SF Object}
\usage{
sf_sf(df_opts = list(), active_column = NULL, active_opts = list(), ...)
}
\arguments{
\item{df_opts}{SF object is a dataframe too,
this is a list where they will be parsed to the Data.frame type}

\item{active_column}{Name of the active column}

\item{active_opts}{List with possible params of sf_sfc to apply
to the active column}

\item{...}{Parsed to assertion_factory}
}
\value{
Assertion for sf
}
\description{
typed for SF object from SF
}
